Add the middleware files
