+++ spring-boot-hello-world/Uploads/File.txt
@@ -1,0 +1,27 @@
+﻿
+Evaluation Only. Created with Aspose.Words. Copyright 2003-2022 Aspose Pty Ltd.
+TITLE
+The phone number format is written like an email
+
+SEVERITY: HIGH
+
+DESCRIPTION
+When a user tries to put his/her number on this field, the webpage will throw an error clarifying that this field needs an @ to be accepted
+
+REQUIREMENTS
+- Have access to https://qa-bugs-hunting-demo-site.herokuapp.com/
+
+STEPS
+1. Open your browser
+2. Click on Add Users Tab
+3. Fill the phone number field with numbers
+
+ACTUAL RESULT
+The Webpage requests the user to put a email format on this field
+
+EXPECTED RESULT
+The Webpage only accpets numbers as Phone NUmber
+
+ADDITIONAL INFORMATION
+Visits this section directly from the url: https://qa-bugs-hunting-demo-site.herokuapp.com/add-user
+Created with an evaluation copy of Aspose.Words. To discover the full versions of our APIs please visit: https://products.aspose.com/words/
