+++ spring-boot-hello-world/src/main/java/com/jalasoft/convert/middleware/FileControllerMiddleware.java
@@ -1,8 +1,13 @@
-package com.jalasoft.convert.controller.middleware;
+package com.jalasoft.convert.middleware;
 
 import com.jalasoft.convert.common.logger.At18Logger;
 import org.apache.poi.EmptyFileException;
+
 import javax.servlet.*;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
 import javax.servlet.annotation.WebFilter;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -10,7 +15,7 @@ import java.io.IOException;
 import java.util.logging.Logger;
 
 /**
- * It is responsible to put a middelware between postman and TranslatorController
+ * It is responsible to put a middelware between postman and FileController
  *
  * Methods to use for HttpServletRequest req: https://docs.oracle.com/javaee/6/api/javax/servlet/ServletRequest.html#getParameterMap()
  * Methods to get the parameters of a file req.getPart("Key file"): https://tomcat.apache.org/tomcat-8.0-doc/api/org/apache/catalina/core/ApplicationPart.html
@@ -18,8 +23,8 @@ import java.util.logging.Logger;
  * @author Mauricio Aliendre
  * @version 1.0
  */
-@WebFilter(urlPatterns = "/uploadGText/*")
-public class TranslatorMiddleware implements Filter {
+@WebFilter(urlPatterns = "/uploadFile")
+public class FileControllerMiddleware implements Filter{
     private static final Logger LOG = new At18Logger().getLogger();
     @Override
     public void doFilter(ServletRequest request,
@@ -29,12 +34,12 @@ public class TranslatorMiddleware implements Filter {
         HttpServletRequest req = (HttpServletRequest) request;
         HttpServletResponse res = (HttpServletResponse) response;
         try {
-            if(((req.getParameter("langI").length() == 2) && req.getParameter("langO").length() == 2) && (req.getPart("text").getContentType().contains("text") || req.getPart("text").getContentType() != null)){
+            if((req.getPart("text").getContentType() != null && res.getStatus() == 200 && req.getPart("file").getSize() > 12000)){
                 LOG.info("Proccess Executed Sucessfully");
                 chain.doFilter(request, response);
-                LOG.info ("Response Status Code is " + res.getStatus());
+                LOG.info ("Response Status Code is: " + res.getStatus());
             } else {
-                LOG.info("the file is no file or there is an empty field");
+                LOG.info("Status is not 200 or the file is Emtpy or there is no file");
                 throw new EmptyFileException();
             }
         } catch (InstantiationError ie){
