+++ spring-boot-hello-world/src/main/java/com/jalasoft/convert/middleware/OcrControllerMiddleware.java
@@ -1,9 +1,13 @@
-package com.jalasoft.convert.controller.middleware;
+package com.jalasoft.convert.middleware;
 
 import com.jalasoft.convert.common.logger.At18Logger;
 import org.apache.poi.EmptyFileException;
 
 import javax.servlet.*;
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
 import javax.servlet.annotation.WebFilter;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -11,7 +15,7 @@ import java.io.IOException;
 import java.util.logging.Logger;
 
 /**
- * It is responsible to put a middelware between postman and ImageController
+ * It is responsible to put a middelware between postman and OcrController
  *
  * Methods to use for HttpServletRequest req: https://docs.oracle.com/javaee/6/api/javax/servlet/ServletRequest.html#getParameterMap()
  * Methods to get the parameters of a file req.getPart("Key file"): https://tomcat.apache.org/tomcat-8.0-doc/api/org/apache/catalina/core/ApplicationPart.html
@@ -19,8 +23,8 @@ import java.util.logging.Logger;
  * @author Mauricio Aliendre
  * @version 1.0
  */
-@WebFilter(urlPatterns = "/uploadImage")
-public class ImageControllerMiddleware implements Filter{
+@WebFilter(urlPatterns = "/uploadOcrImg")
+public class OcrControllerMiddleware implements Filter{
     private static final Logger LOG = new At18Logger().getLogger();
     @Override
     public void doFilter(ServletRequest request,
@@ -30,7 +34,7 @@ public class ImageControllerMiddleware implements Filter{
         HttpServletRequest req = (HttpServletRequest) request;
         HttpServletResponse res = (HttpServletResponse) response;
         try {
-            if((req.getPart("text").getContentType() != null && res.getStatus() == 200)){
+            if(req.getPart("text").getContentType() != null && res.getStatus() == 200){
                 LOG.info("Proccess Executed Sucessfully");
                 chain.doFilter(request, response);
                 LOG.info ("Response Status Code is: " + res.getStatus());
