+++ spring-boot-hello-world/src/main/java/com/jalasoft/convert/middleware/VideoControllerMiddleware.java
@@ -1,8 +1,7 @@
-package com.jalasoft.convert.controller.middleware;
+package com.jalasoft.convert.middleware;
 
 import com.jalasoft.convert.common.logger.At18Logger;
 import org.apache.poi.EmptyFileException;
-
 import javax.servlet.*;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -15,7 +14,7 @@ import java.io.IOException;
 import java.util.logging.Logger;
 
 /**
- * It is responsible to put a middelware between postman and audiocontroller
+ * It is responsible to put a middelware between postman and VideoController
  *
  * Methods to use for HttpServletRequest req: https://docs.oracle.com/javaee/6/api/javax/servlet/ServletRequest.html#getParameterMap()
  * Methods to get the parameters of a file req.getPart("Key file"): https://tomcat.apache.org/tomcat-8.0-doc/api/org/apache/catalina/core/ApplicationPart.html
@@ -23,29 +22,29 @@ import java.util.logging.Logger;
  * @author Mauricio Aliendre
  * @version 1.0
  */
-@WebFilter(urlPatterns = "/uploadAudio")
-public class AudioControllerMiddleware implements  Filter{
+@WebFilter(urlPatterns = "/uploadVideo")
+public class VideoControllerMiddleware implements Filter{
     private static final Logger LOG = new At18Logger().getLogger();
-
     @Override
     public void doFilter(ServletRequest request,
                          ServletResponse response,
-                         FilterChain chain) throws IOException, ServletException {
+                         FilterChain chain) throws IOException, ServletException
+    {
         HttpServletRequest req = (HttpServletRequest) request;
         HttpServletResponse res = (HttpServletResponse) response;
         try {
-            if ((req.getPart("text").getContentType() != null && res.getStatus() == 200 && req.getPart("file").getSize() > 12000)) {
+            if(req.getPart("text").getContentType() != null && res.getStatus() == 200){
                 LOG.info("Proccess Executed Sucessfully");
                 chain.doFilter(request, response);
-                LOG.info("Response Status Code is: " + res.getStatus());
+                LOG.info ("Response Status Code is: " + res.getStatus());
             } else {
-                LOG.info("Status is not 200 or the file is Emtpy or there is no file");
+                LOG.info("the file is Emtpy or there is a blank field");
                 throw new EmptyFileException();
             }
-        } catch (InstantiationError ie) {
+        } catch (InstantiationError ie){
             LOG.info("Catch Instantiation Error: " + ie);
             ie.printStackTrace();
-        } catch (NullPointerException nulle) {
+        } catch (NullPointerException nulle){
             LOG.info("Catch a null pointer exception: " + nulle);
             nulle.printStackTrace();
         }
