+++ spring-boot-hello-world/traducido -File.txt
@@ -1,0 +1,27 @@
+﻿
+Solo evaluación. Creado con Aspose.Words. Copyright 2003-2022 Aspose Pty Ltd.
+TÍTULO
+El formato del número de teléfono se escribe como un correo electrónico.
+
+GRAVEDAD: ALTA
+
+DESCRIPCIÓN
+Cuando un usuario intente poner su número en este campo, la página web arrojará un error aclarando que este campo necesita una @ para ser aceptado
+
+REQUISITOS
+- Tener acceso a https://qa-bugs-hunting-demo-site.herokuapp.com/
+
+PASOS
+1. Abre tu navegador
+2. Haga clic en la pestaña Agregar usuarios
+3. Complete el campo de número de teléfono con números
+
+RESULTADO ACTUAL
+La página web solicita al usuario que coloque un formato de correo electrónico en este campo
+
+RESULTADO ESPERADO
+La página web solo acepta números como NÚMERO DE TELÉFONO
+
+INFORMACIÓN ADICIONAL
+Visita esta sección directamente desde la URL: https://qa-bugs-hunting-demo-site.herokuapp.com/add-user
+Creado con una copia de evaluación de Aspose.Words. Para descubrir las versiones completas de nuestras API, visite: https://products.aspose.com/words/
